Interface People et TEAMSProcessor, sans test pour l'instant
