+++ src/appli/People.java
@@ -6,118 +6,33 @@ import java.time.format.DateTimeFormatter;
 import java.util.LinkedList;
 import java.util.Locale;
 
-public abstract class People implements Comparable<People> {
+public interface People extends Comparable<People> {
 
-    protected final String _name;
-    protected final String _id;
-    protected String _start;
-    protected String _stop;
-    protected LinkedList<TEAMSPeriod> _periodList;
 
-    public People(String _name) {
-        this._name = _name;
-        this._id = StudentIDServer.getId(this._name);
-        this._periodList = new LinkedList<>();
-    }
+    public String get_id();
 
-    public String get_id() {
-        return _id;
-    }
+    public void addPeriod(String action, String instant);
 
-    public void addPeriod(String action, String instant) {
+    public void forceEndTimeAt(String instant);
 
-        if ( action.charAt(0) == 'R' ) {
-            TEAMSPeriod period = new TEAMSPeriod(instant);
-            this._periodList.add(period);
-        } else
-            if ( action.charAt(0) == 'A' ) {
-                this._periodList.getLast().stopAt(instant);
-            } else {
-                System.out.println(this._name + " --> erreur : action inconnue ["+action+"] ");
-            }
-    }
+    public void forceStartTimeAt(String instant);
 
-    public void forceEndTimeAt(String instant) {
-        this._stop = instant;
-        // delete all periods started after ending time
-        var localPeriodEndingTime = TEAMSDateTimeConverter.StringToLocalDateTime(this._stop);
-        boolean STOP = false;
-        while (!STOP) {
-            if (this._periodList.size()>0 && this._periodList.getLast().get_start().isAfter(localPeriodEndingTime))
-                this._periodList.removeLast();
-            else
-                STOP = true;
-        }
+    public long getTotalAttendanceDuration();
 
-        // no ending time, so fix it
-        if (this._periodList.size()>0 && this._periodList.getLast().get_end()==null)
-            this._periodList.getLast().stopAt(this._stop);
-        else
-        // ending after the official end of the course, so limit it
-        if (this._periodList.size()>0 && this._periodList.getLast().get_end().isAfter(localPeriodEndingTime))
-            this._periodList.getLast().stopAt(this._stop);
-    }
+    public boolean isClosed();
 
-    public void forceStartTimeAt(String instant) {
-        this._start = instant;
-        // delete all periods ended before starting time
-        var periods = this._periodList.iterator();
-        boolean STOP = false;
-        var localPeriodStartingTime = TEAMSDateTimeConverter.StringToLocalDateTime(this._start);
-        while ( periods.hasNext() && !STOP) {
-            var period = periods.next();
-            if (period.get_end().isBefore(localPeriodStartingTime))
-                periods.remove();
-            else
-                STOP = true;
-        }
-        // adjust starting time of period if before official course starting time
-        if (!this._periodList.isEmpty() && this._periodList.getFirst().get_start().isBefore(localPeriodStartingTime))
-            this._periodList.getFirst().startAt(this._start);
+    public String getName();
 
-    }
+    public String getDate();
 
-    public long getTotalAttendanceDuration() {
-        double totalDuration = 0;
-        for (TEAMSPeriod period : this._periodList) {
-            totalDuration += period.getDurationInMinutes();
-        }
-        return Math.round(totalDuration);
-    }
+    public boolean isOutOfPeriod();
 
-    public boolean isClosed() {
-        return this._periodList.getLast().isEnded();
-    }
-
-    public String getName() {
-        return this._name;
-    }
-
-    public abstract String getCode();
+    public String getCode();
 
     @Override
-    public String toString() {
-        return "People{" +
-                "_name='" + _name + '\'' +
-                ", _id='" + _id + '\'' +
-                ", _periodList=" + _periodList +
-                ", _start='" + _start + '\'' +
-                ", _stop='" + _stop + '\'' +
-                '}';
-    }
-
-    public String getDate() {
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd MMMM yyyy");
-        return ( this._periodList.getFirst().get_start().format(formatter.withLocale(Locale.FRANCE)) );
-    }
+    public String toString();
 
     @Override
-    public int compareTo(People o) {
-        return (int)(this.getTotalAttendanceDuration()-o.getTotalAttendanceDuration());
-    }
-
-    public boolean isOutOfPeriod() {
-        return this._periodList.isEmpty();
-    }
+    public int compareTo(People o);
 
 }
