+++ src/appli/TEAMSProcessor.java
@@ -6,44 +6,11 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
-public abstract class TEAMSProcessor {
+public interface TEAMSProcessor {
 
-    protected Collection<People> _allpeople = null;
-    protected String _fileName;
-    protected String _startTime;
-    protected String _endTime;
+    Collection<People> _allpeople = null;
 
-    public TEAMSProcessor(File _file, String _start, String _stop) {
-        /*
-         csv file to read
-         start time of the course
-         end time of the source
-        */
-        this._startTime = _start;
-        this._endTime = _stop;
+    public Collection<People> get_allpeople();
 
-        // load CSV file
-        this._fileName = _file.getName();
-        var teamsFile = new TEAMSAttendanceList(_file);
-
-        // filter to extract data for each people
-        var lines = teamsFile.get_attlist();
-        if (lines != null) {
-            // convert lines in data structure with people & periods
-            var filter = new TEAMSAttendanceListAnalyzer(lines);
-            // cut periods before start time and after end time
-            filter.setStartAndStop(_start, _stop);
-            // sort
-            List<People> peopleByDuration = new ArrayList<>(filter.get_peopleList().values());
-            Collections.sort(peopleByDuration);
-            // init the people collection
-            this._allpeople = peopleByDuration;//filter.get_peopleList().values();
-        }
-    }
-
-    public Collection<People> get_allpeople() {
-        return _allpeople;
-    }
-
-    public abstract String codeTo();
+    public String codeTo();
 }
