+++ src/appli/TEAMSProcessorHTML.java
@@ -1,11 +1,49 @@
 package appli;
 
 import java.io.File;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
 
-public class TEAMSProcessorHTML extends TEAMSProcessor{
+public class TEAMSProcessorHTML implements TEAMSProcessor{
+
+    private Collection<People>_allpeople = null;
+    private String _fileName;
+    private String _startTime;
+    private String _endTime;
 
     public TEAMSProcessorHTML(File _file, String _start, String _stop) {
-        super(_file, _start, _stop);
+            /*
+         csv file to read
+         start time of the course
+         end time of the source
+        */
+        this._startTime = _start;
+        this._endTime = _stop;
+
+        // load CSV file
+        this._fileName = _file.getName();
+        var teamsFile = new TEAMSAttendanceList(_file);
+
+        // filter to extract data for each people
+        var lines = teamsFile.get_attlist();
+        if (lines != null) {
+            // convert lines in data structure with people & periods
+            var filter = new TEAMSAttendanceListAnalyzer(lines);
+            // cut periods before start time and after end time
+            filter.setStartAndStop(_start, _stop);
+            // sort
+            List<People> peopleByDuration = new ArrayList<>(filter.get_peopleList().values());
+            Collections.sort(peopleByDuration);
+            // init the people collection
+            this._allpeople = peopleByDuration;//filter.get_peopleList().values();
+        }
+    }
+
+    @Override
+    public Collection<People> get_allpeople() {
+        return null;
     }
 
     public String codeTo() {
